E4X -> DOM rewrite
E4X is a cumbersome specification which implements a lot of unintuitive
behavior. The goal of e4x-bump was primarily focused on bringing XML
literals to Javascript. Since prior work existed in this space it was
decided to work within the E4X spec, but it seems that a more DOM-like
interface will be cleaner in the end.

So, this is the beginning of a full runtime rewrite which will work with
a virtual DOM-like structure. I'm implementing a library loosely based
on DOM level 2, including namespaces. This should have a lot of
advantages down the road, including simpler server code, faster client
code, and better transparency for existing DOM developers.