HTML Support
This commit adds definitions for HTML elements. This is important
because it gets attributes into their own object by defining explicit
accessors for each attribute.

This also implements the notion of an XMLEnvironment. It's difficult to
compartmentalize the behavior of XML literals because they are
inherently global. So I wrapped up all flags which can effect an XML
literal expression into an XMLEnvironment which you can configure and
reuse.