Add support for unescaped data
It's likely that due to legacy systems or 3rd parties that clients will
need to inject raw HTML into their pages, bypassing the safety that pure
XML literal documents gets them. While unfortunate, it's a real world
need. This adds a `UnsafeUnescapeData` constructor which, when rendered,
will not espace its data.

This also fixes several bugs in the runtime with rendering and
construction.