Abstract out interface for XML literals
This diff moves the construction of XML literals into XMLEnvironment
which lets you create different environments for different purposes. For
instance you could have one environment which operates on a client-side
DOM, or you could have another environment which creates a simulated
server-side DOM. The choice is up to the user. Since this project is
focusing on XML literals only, it makes sense to decouple from the
implementation of the DOM.