Native DOM environment
Adds an environment which supports directly manipulating existing DOM
implementations. There's a new `DOMXMLEnvironment` constructor which
only needs a `document` object and will enable full support for XML
literals. You can use this on the client-side with the browser's DOM, or
on the server-side using jsdom (or other DOM libraries).